+++ packages/dash-table/src/dash-table/derived/cell/eventHandler.ts
@@ -1,12 +1,15 @@
 
 import valueCache from 'core/cache/value';
 import { ICellFactoryProps } from 'dash-table/components/Table/props';
-import { handleChange, handleClick, handleDoubleClick, handleOnMouseUp, handlePaste } from 'dash-table/handlers/cellEvents';
+import { handleChange, handleClick, handleDoubleClick, handleEnter, handleLeave, handleMove, handleOnMouseUp, handlePaste } from 'dash-table/handlers/cellEvents';
 
 export enum Handler {
     Change = 'change',
     Click = 'click',
     DoubleClick = 'doubleclick',
+    Enter = 'enter',
+    Leave = 'leave',
+    Move = 'move',
     MouseUp = 'mouseup',
     Paste = 'paste'
 }
@@ -30,6 +33,12 @@ class EventHandler {
                 return handleClick.bind(undefined, this.propsFn, rowIndex, columnIndex);
             case Handler.DoubleClick:
                 return handleDoubleClick.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Enter:
+                return handleEnter.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Leave:
+                return handleLeave.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Move:
+                return handleMove.bind(undefined, this.propsFn, rowIndex, columnIndex);
             case Handler.MouseUp:
                 return handleOnMouseUp.bind(undefined, this.propsFn, rowIndex, columnIndex);
             case Handler.Paste: