+++ packages/dash-table/src/dash-table/derived/cell/eventHandlerProps.ts
@@ -12,6 +12,9 @@ interface IFunctionCache {
     onChange: (e: ChangeEvent) => void;
     onClick: (e: MouseEvent) => void;
     onDoubleClick: (e: MouseEvent) => void;
+    onEnter: (e: MouseEvent) => void;
+    onLeave: (e: MouseEvent) => void;
+    onMove: (e: MouseEvent) => void;
     onMouseUp: (e: MouseEvent) => void;
     onPaste: (e: ClipboardEvent<Element>) => void;
 }
@@ -24,6 +27,9 @@ const getter = (propsFn: () => ICellFactoryProps) => {
             onChange: derivedHandlers(Handler.Change, columnIndex, rowIndex),
             onClick: derivedHandlers(Handler.Click, columnIndex, rowIndex),
             onDoubleClick: derivedHandlers(Handler.DoubleClick, columnIndex, rowIndex),
+            onEnter: derivedHandlers(Handler.Enter, columnIndex, rowIndex),
+            onLeave: derivedHandlers(Handler.Leave, columnIndex, rowIndex),
+            onMove: derivedHandlers(Handler.Move, columnIndex, rowIndex),
             onMouseUp: derivedHandlers(Handler.MouseUp, columnIndex, rowIndex),
             onPaste: derivedHandlers(Handler.Paste, columnIndex, rowIndex)
         };