+++ packages/dash-table/tests/cypress/plugins/index.js
@@ -6,4 +6,13 @@ module.exports = on => {
     };
 
     on('file:preprocessor', wp(options));
+
+    on('before:browser:launch', (browser = {}, args) => {
+        if (browser.name === 'chrome') {
+            // ^ make sure this is your browser name, you may
+            // be using 'canary' or 'chromium' for example, so change it to match!
+            args.push('--proxy-bypass-list=<-loopback>');
+            return args;
+        }
+    });
 };