+++ packages/dash-table/tests/visual/percy-storybook/@Types/modules.d.ts
@@ -15,3 +15,12 @@ declare module 'fast-isnumeric' {
 
     export default value;
 }
+
+declare class Remarkable {
+    constructor();
+    render(value: string): any;
+}
+
+declare module 'remarkable' {
+    export default Remarkable;
+}